Fixed a lot of warnings

+++ source/user/c/time/timemain.c
@@ -59,7 +59,7 @@ int main(int argc,char **argv) {
 		const char **args = (const char**)malloc(sizeof(char*) * (argc - 1));
 		if(!args)
 			error("Not enough mem");
-		for(i = 1; i < argc; i++)
+		for(i = 1; i < (size_t)argc; i++)
 			args[i - 1] = argv[i];
 		args[argc - 1] = NULL;
 		execp(args[0],args);
